Split up authorization router
